+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- 2021-01-07 Add `TimeDuration`.
+- 2021-01-07 Add `TimeDuration`. Add a lot of operator overloads for `TimeDuration` and `TimeTick`.
 - 2021-12-31 Add default constructor for `Some<T>`.
 - 2021-12-10 Improve JsonNode. Remove `Some<JsonNode>` etc. |api-break|